Split up absl into many sub-libraries
When absl is distributed in the open-source world, both in its bazel
and cmake files, it's distributed as a bunch of little libraries.
However, we were building just one mega absl library. The mega library
causes problems with the absl::flags library, as that has global
variables that fail at runtime if they're included twice. When adding
a dep from protobuf -> absl, we were hitting those errors. However,
protobuf doesn't actually require absl::flags, so we need smaller build
targets so that protobuf can just depend on what it needs.

Autogenerate all the small library targets from the bazel build file
definitions.

Usages of the mega library will be switched to the appropriate
sub-libraries after this is submitted.

Bug: 417751597
Test: full build on aosp_cf_x86_64_only_phone
Change-Id: I7f32a5501b19e19c12b4174c5bd8c87967343f57